requirements.txt
# Core dependencies
PyYAML>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.1
loguru>=0.7.2

# NLP and Intelligence
transformers>=4.35.2
torch>=2.1.1
numpy>=1.26.2
scikit-learn>=1.3.2
sentence-transformers>=2.2.2
nltk>=3.8.1
spacy>=3.7.2
textblob>=0.17.1
vaderSentiment>=3.3.2

# Voice Processing
openai-whisper>=20231117
SpeechRecognition>=3.10.0
pydub>=0.25.1
pyaudio>=0.2.13
sounddevice>=0.4.6
librosa>=0.10.1
coqui-tts>=0.21.0
pyvoice>=0.1.4
pvporcupine>=3.0.0  # Picovoice wake word

# Computer Vision
opencv-python>=4.8.1.78
dlib>=19.24.2
face-recognition>=1.3.0
Pillow>=10.1.0
mediapipe>=0.10.8

# Storage and Memory
redis>=5.0.1
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
chromadb>=0.4.18
hnswlib>=0.7.0

# Web & Interfaces
fastapi>=0.104.1
uvicorn>=0.24.0.post1
websockets>=12.0
requests>=2.31.0
flask>=3.0.0
pyngrok>=7.0.0  # For tunneling during development

# Security
cryptography>=41.0.5
bcrypt>=4.0.1
python-jose>=3.3.0
passlib>=1.7.4

# Smart Home Integration
paho-mqtt>=2.2.1
python-homeassistant-api>=1.0.0
roku>=4.2.0
pysmartthings>=0.7.8

# Utilities
colorlog>=6.7.0
click>=8.1.7
schedule>=1.2.1
python-dateutil>=2.8.2

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Android Communication
pyjnius>=1.6.1  # For Python-Android communication

